load_N_ask/requirements.txt
langchain
faiss-cpu
pypdf2
streamlit-chat
requests
openai
tiktoken
